Bug fix #2659
The bugfix for linking recursively from files worked, but there was another
bug hidden... only showed up when actually *using* the recursive feature,
and not with only testing the bugfix. :) Famous coder error, but yah, we
got luckily users to test it!

I'v asked the reporter to give it more extensive testing!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4548 954f8c5b-7b00-dc11-b283-0030488c597c