Boundbox select object mode optimisation.
Changed selectprojektie (and renamed to view3d_opengl_select as suggested by Ton) to accept a buffer size, so boundbox can adapt it's buffer size to the number of object in scene.
Also, the loop is done more smartly, since glSelect fills the buffer in the same order as the drawing order, so we save lops of looping on unselected object (which used to go through all the buffer before finding that they weren't selected).

This scheme could probably be applied to all the other loops using glSelect. (good project for newbie coders)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4686 954f8c5b-7b00-dc11-b283-0030488c597c