While checking old bugreports on the todo tracker, found weak issue in current depgraph... that's when an object has multiple "parents", which have different relation types. (one parent moves Object, the other deforms). In that case the result was undefined.
Solved it nicely with the DAG building code, resulting in quicker flushing
of changes even.
Also noted a bug in the drawview.c 'object_handle_update' call, which
should not check for layers. The DAG flushing clears recalc flags for
invisible layers, if possible. Now need to write that @#!# doc!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4781 954f8c5b-7b00-dc11-b283-0030488c597c