- shuffled editmesh derived function name/function  - added ModifierTypeInfo.freeData function  - added modifier_{new,free] utility function  - added ccgSubSurf_getUseAgeCounts to query info  - removed subsurf modifier faking (ME_SUBSURF flag is no    longer valid). subsurf modifier gets converted on file load    although there is obscure linked mesh situation where this    can go wrong, will fix shortly. this also means that some    places in the code that test/copy subsurf settings are broken    for the time being.  - shuffled modifier calculation to be simpler. note that    all modifiers are currently disabled in editmode (including    subsurf). don't worry, will return shortly.  - bug fix, build modifier didn't randomize meshes with only verts  - cleaned up subsurf_ccg and adapted for future editmode modifier    work  - added editmesh.derived{Cage,Final}, not used yet  - added SubsurfModifierData.{mCache,emCache}, will be used to cache    subsurf instead of caching in derivedmesh itself  - removed old subsurf buttons  - added do_modifiers_buttons to handle modifier events  - removed count_object counting of modifier (subsurfed) objects...    this would be nice to add back at some point but requires care.    probably requires rewrite of counting system.
New feature: Incremental Subsurf in Object Mode

The previous release introduce incremental subsurf calculation during
editmode but it was not turned on during object mode. In general it
does not make sense to have it always enabled during object mode because
it requires caching a fair amount of information about the mesh which
is a waste of memory unless the mesh is often recalculated.

However, for mesh's that have subsurfed armatures for example, or that
have other modifiers so that the mesh is essentially changing on every
frame, it makes a lot of sense to keep the subsurf'd object around and
that is what the new incremental subsurf modifier toggle is for. The
intent is that the user will enable this option for (a) a mesh that is
currently under active editing or (b) a mesh that is heavily updated
in the scene, such as a character.

I will try to write more about this feature for release, because it
has advantages and disadvantages that are not immediately obvious (the
first user reaction will be to turn it on for ever object, which is
probably not correct).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4865 954f8c5b-7b00-dc11-b283-0030488c597c