Modifications for bullet were made and variables were defined only for windows.  This commit defines these variables for the other blender platforms helping SCons not to barf quite as quickly.  Scons is still unable to compile a fresh CVS checkout at the present time on linux, but now at least allows config.opts to be properly created.
The new defines for use_bullet are currently set as false until testing can take place to ensure proper compilation on those systems.  The existing use_bullet='true' on the windows platform was left untouched.

Please ensure that variables are defined on _all_ platforms in the future instead of just defining them for the system you're currently developing under.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4993 954f8c5b-7b00-dc11-b283-0030488c597c