Roland Hess' Floor Constraint patch: https://projects.blender.org/tracker/?func=detail&aid=2993&group_id=9&atid=127
Minor modifications to simplify the code in evaluate_constraint.

The "Stick" feature will need more work as it gives bad results when skipping frames, jumping around on the timeline and when going backward in time.

Suggestion: Would be nice if it could use the local space too, not just global space planes.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@5246 954f8c5b-7b00-dc11-b283-0030488c597c