This is a bit a patch... unforseen issue with the new Screen Handlers. (Screen handler is used now to trigger animated screen).
When files get saved with Verse_Blender, it has the verse handler set,
which causes the handler system to run without a means to disable it with
release versions of Blender. This patch disables the Verse Handler to be
set on reading files.

Jiri; you will have to uncomment this code when you sync it with your tree!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@5598 954f8c5b-7b00-dc11-b283-0030488c597c