Previewrender (buttons) now supports Lamp, Texture and Sky previews again. One exception I didn't solve yet... for the Stucci texture.
In a next commit I will add the preview.blend, so people can play with it!

The rules for preview.blend are pretty primitive still... might become
a bit more advanced later. Here's what it uses now:

- If Object has name starting with 'p' it will get the to-be-previewed
  Material or Lamp assigned
- If Object has name starting with 't' its Material gets the texture
  assigned for Texture previews

The layer codes for current preview options are in DNA_material_types.h;

#define MA_SPHERE		0
#define MA_CUBE			1
#define MA_FLAT			2
#define MA_MONKEY		3
#define MA_SPHERE_A		4
#define MA_TEXTURE		5
#define MA_LAMP			6
#define MA_SKY			7

(SPHERE_A is sphere with alpha for icons)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@6581 954f8c5b-7b00-dc11-b283-0030488c597c