Fix in blur filter 'bokeh'.
This filter type uses a filter-image, and spreads color of current pixel
over all neighbour pixels based on this filter-image. That creates a
problem on borders... since there only parts get accumulated.

Solved by going over the to-be-filtered image pretending it is wider
exactly the amount of the filter-image size.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@6814 954f8c5b-7b00-dc11-b283-0030488c597c