Bugfix #4066
Particle strands with a width set (like 10 pixels), and with extreme
bending of strands (like a very course subdivision), could create non-flat
quads. For speed reasons, the quad-to-triangle splitting was turned off
for hair, but in this case that should be done nevertheless.

Solves another Dandruff issue for furry bunnies!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7224 954f8c5b-7b00-dc11-b283-0030488c597c