Bugfix #4420 + #3975
Found reasonable OK solution for a cyclic dependency that was bothering
a lot of (character) animators. They would like to drive bones with
Objects, but that lagged when you parent these Objects to the
armature, since that's a circular dependency.

Since Driver relations are only looking at local object properties (i.e.
it's local position/rotation) there's a simple way to solve it. In case
such Objects have an Ipo, it reads driver values directly from the
Object Ipo (on correct time) instead of Object values.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7852 954f8c5b-7b00-dc11-b283-0030488c597c