Bugfix #4530
Weightpaint fix.

Weight painting was recoded in 2.40 to support all features from vertex
paint. That code uses a 'soft' brush by default, which makes it impossible
to assign exact weight values on a single click (which used to be possible)

I've made that an option now, so you can disable it for single-click setting
of values on vertices within the brush (if 'use vertex distances' is
set, of course).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7977 954f8c5b-7b00-dc11-b283-0030488c597c