Another fix based on venomgfx fraka .blend:
In december, when testing material layering, I've removed the check that
prevented specular and diffuse to become negative, this because it could
work nice for layering.

However, this breaks quite some cases too. For example negative lamps are
only used to cancel out other lights in same material, and should not give
negative (invisible!) peaks that work on a node system. Same goes for
negative diffuse from AO 'subtract' mode.

In fraka the error happened for AO on a translucent material. The inside
of the mesh got a negative AO, cancelling out the positive AO on the
outside.

Anyhoo; this commits ensures that a 'shade_lamp_loop' call will never
return negative values again!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8110 954f8c5b-7b00-dc11-b283-0030488c597c