Bugfix: (own todo)
When you rename a generated image, like "Render Result" or "Composite"
the float buffer in this image had to be copied, otherwise a next render
will make the renamed Image buffer invalid.

Note that renaming will not mean the images get packed or saved! So on
reloading .blend they're gone.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8115 954f8c5b-7b00-dc11-b283-0030488c597c