Bugfix #4773
The VectorBlur node crashed when it didn't get a RGBA type image as
input (like using Alpha as input). Added typeconversion for it.

Note: the Z input and Vector input are not converted, but checked for.
When an improper type gets connected to these inputs, an error is
printed in console.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8204 954f8c5b-7b00-dc11-b283-0030488c597c