Bugfix 4865
Manipulator orientation tweak: when having multiple objects selected, the
orientation option "Local" was always skipped, showing it global. Now it
uses the orientation of the active object.

The bug reportor noticed a conflict with the "Around" option "Active". In
that case he expected the orientation to be from Active Object as well.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8306 954f8c5b-7b00-dc11-b283-0030488c597c